Use subclass in TypeScript
